Fix test_window_size on windows.
On Windows, there seems to be a minimum size to the client window allowed,
likely due to the size of the title bar. Since we're only checking to see
that the dimensions were properly updated, we can use different dimensions
instead of trying to find failing edge cases (that's a GLFW problem).