Hide Vulkan functions behind IncludeVulkan cabal flag.
This mimics the way that the GLFW interface prevents access to vulkan functions
without user intervention. In particular, the user of GLFW must specify some
Vulkan header in order to get access to those functions, either by including a
header prior to glfw3.h or by using -DGLFW_INCLUDE_VULKAN. Here, we can only
really do the latter, so we expect clients of this library to use the
IncludeVulkan flag and the appropriate extra-include-dirs stack/cabal flag when
they want to use the associated Vulkan interface.

To prevent incompatibilities between API versions, we also stub out the would-be
generated functions with placeholders that error when they're called,
instructing the user to enable the flag if they'd like access to those
functions.

In the tests, we don't really have the resources to set up an entire Vulkan
instance. Instead, simply make sure that we get the proper error values out
of the functions that we're dealing with, and that we can call them regardless
of the IncludeVulkan flag.