On second thought, always keep Vulkan functions enabled.
Left the comment in there to warn potential users. However, the M.O. of most
Haskell libraries seems to be to lower overhead of use as much as possible. This
means that we should simply assume that if the user is using the Vulkan
functions, they probably have Vulkan bindings that will generate the appropriate
types and will generally know how to use these functions. To reduce friction, we
should probably not make them go fishing for include files. This (in my opinion)
isn't the "right" way to do this, but it probably is the one that will be the
nicest for clients of this library.