Time val and frequency should be Word64s not CULong.
(Unconfirmed) I think what happens on Windows when you try to marshall from a
non standards compliant type, such as one from stdint.h, ghc doesn't know the
proper type to convert it to. Therefore, when we're going from uint64_t to
CULong, ghc first converts it to a 32-bit representation, because that's the
only thing it has guarantees for. We have to explicitly specify that the return
value here is 64 bits in order to get proper marshalling. This same problem
doesn't seem to manifest itself on other platforms, such as OS X.